+++ src/components/modules/Home/ProjectCard/style.css
@@ -55,25 +55,3 @@
   background-color: var(--primary-400);
   border-radius: 5rem;
 }
-
-.delete-confirm-modal {
-  position: absolute;
-  left: 50%;
-  transform: translateX(-50%);
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-  box-shadow: 0px 0px 8px -2px rgba(89, 142, 220, 0.54);
-  background-color: var(--grey-100);
-  padding: 2rem;
-  border-radius: 25px;
-}
-
-.modal-display {
-  display: flex;
-  z-index: 999;
-}
-
-.modal-hidden {
-  display: none;
-}
